refactor: move canvas to canvas/ and decompose viewport by volatility
Lowy analysis found useCanvasViewport conflated three independent
volatility axes. Decomposed into:

  canvas/viewport/gestures.ts    — wheel + middle-mouse event handling
  canvas/viewport/transforms.ts  — pure math (fitAll, zoomToCenter, snap)
  canvas/viewport/coordinates.ts — CSS generation (transform, grid bg)
  canvas/viewport/useCanvasViewport.ts — orchestrator wiring the three

Also moved all canvas code from terminal/ to canvas/ and extracted
CanvasTile into its own file.
