+++ packages/client/src/App.tsx
@@ -15,7 +15,7 @@ import Header from "./Header";
 import PwaInstallBar from "./PwaInstallBar";
 import Sidebar from "./sidebar/Sidebar";
 import TerminalContent from "./terminal/TerminalContent";
-import TerminalCanvas from "./terminal/TerminalCanvas";
+import TerminalCanvas from "./canvas/TerminalCanvas";
 import MobileKeyBar from "./MobileKeyBar";
 import CommandPalette from "./CommandPalette";
 import ShortcutsHelp from "./ShortcutsHelp";
@@ -35,8 +35,8 @@ import { useThemeManager } from "./useThemeManager";
 import { useSidebar } from "./sidebar/useSidebar";
 import { useShortcuts } from "./input/useShortcuts";
 import { useSubPanel } from "./terminal/useSubPanel";
-import { useCanvasViewport } from "./terminal/useCanvasViewport";
-import { useCanvasLayouts, type TileLayout } from "./terminal/useCanvasLayouts";
+import { useCanvasViewport } from "./canvas/viewport/useCanvasViewport";
+import { useCanvasLayouts, type TileLayout } from "./canvas/useCanvasLayouts";
 import { useRightPanel } from "./right-panel/useRightPanel";
 import { useColorScheme } from "./settings/useColorScheme";
 import { useServerState } from "./settings/useServerState";
