+++ packages/client/src/terminal/useSessionRestore.ts
@@ -3,7 +3,7 @@
 import { createSignal, createEffect } from "solid-js";
 import { toast } from "solid-sonner";
 import { useSubPanel } from "./useSubPanel";
-import { useCanvasLayouts } from "./useCanvasLayouts";
+import { useCanvasLayouts } from "../canvas/useCanvasLayouts";
 import { useServerState } from "../settings/useServerState";
 import { lifecycle } from "../rpc/rpc";
 import type { TerminalId, TerminalInfo, SavedSession } from "kolu-common";
